skeleton for local all and minor changes

+++ SmartFarmer.Core/Tasks/Movement/IFarmerMoveOnGridTask.cs
@@ -6,6 +6,6 @@ namespace SmartFarmer.Tasks.Movement;
 
 public interface IFarmerMoveOnGridTask : IFarmerTask
 {
-    Task MoveToPosition(int x, int y, CancellationToken token);
-    void GetCurrentPosition(out int x, out int y);
+    Task MoveToPosition(double x, double y, CancellationToken token);
+    void GetCurrentPosition(out double x, out double y);
 }
