+++ SmartFarmer.Tasks.Mock/Movement/FarmerMoveOnGridTask.cs
@@ -9,7 +9,7 @@ namespace SmartFarmer.Tasks.Movement
 {
     public class FarmerMoveOnGridTask : FarmerBaseTask, IFarmerMoveOnGridTask
     {
-        private int _currentX, _currentY;
+        private double _currentX, _currentY;
 
         public FarmerMoveOnGridTask()
         {
@@ -20,13 +20,13 @@ namespace SmartFarmer.Tasks.Movement
         {
             if (parameters == null || parameters.Length < 2) throw new ArgumentException(nameof(parameters));
 
-            var x = (int)parameters[0];
-            var y = (int)parameters[1];
+            var x = (double)parameters[0];
+            var y = (double)parameters[1];
 
             await MoveToPosition(x, y, token);
         }
 
-        public async Task MoveToPosition(int x, int y, CancellationToken token)
+        public async Task MoveToPosition(double x, double y, CancellationToken token)
         {
             PrepareTask();
 
@@ -41,7 +41,7 @@ namespace SmartFarmer.Tasks.Movement
             await Task.CompletedTask;
         }
             
-        public void GetCurrentPosition(out int x, out int y)
+        public void GetCurrentPosition(out double x, out double y)
         {
             x = _currentX;
             y = _currentY;
